Rename SchwerScripts folder to schwer-scripts
