+++ README.md
@@ -10,20 +10,20 @@ This repository contains skeleton implementations of various features. Feel free
 
 I'm just a hobbyist, so if you have any improvements or suggestions, let me know!
 ## Contents
-* [Item System](/SchwerScripts/ItemSystem) (serializable!)
-* [Editor scripts](/SchwerScripts/Editor)
+* [Item System](/schwer-scripts/ItemSystem) (serializable!)
+* [Editor scripts](/schwer-scripts/Editor)
     * `PrefabMenu` (menu items to speed up prefab workflow)
     * `AssetsUtility` (work with assets via code)
     * `ScriptableObjectUtility` (work with Scriptable Object assets via code)
     * TBA
-* [Common scripts](/SchwerScripts/Common)
+* [Common scripts](/schwer-scripts/Common)
     * `BinaryIO` (wrapper for reading and writing binary files)
     * `MonoBehaviourSingleton` & `DDOLSingleton`
 * TBA (WebGL save/loading!)
 
 <br/>
 
-***The rest of this readme is copied from the [ItemSystem readme](/SchwerScripts/ItemSystem/README.md)!***
+***The rest of this readme is copied from the [ItemSystem readme](/schwer-scripts/ItemSystem/README.md)!***
 # schwer-scripts: item system
 A skeleton implementation of items and (serializable!) inventories using Scriptable Objects, as well as several supporting editor tools.
 
@@ -51,8 +51,8 @@ A skeleton implementation of items and (serializable!) inventories using Scripta
 * As `Inventory` assets are Scriptable Objects, there must be a reference to each `Inventory` asset at all times during runtime in order for their values to persist, or Unity will unload them the next time it unloads unused resources.
 
 ## Getting started
-Download this repository and add the `SchwerScripts` folder to your Unity project.
-Familiarise yourself with the system by dragging the prefab `Inventory` (`SchwerScripts/ItemSystem/Demo/UI/Prefabs/Inventory`) into a new Scene and following along with this section.
+Download this repository and add the `schwer-scripts` folder to your Unity project.
+Familiarise yourself with the system by dragging the prefab `Inventory` (`schwer-scripts/ItemSystem/Demo/UI/Prefabs/Inventory`) into a new Scene and following along with this section.
 
 You may need to create an Event System (`Right-click in the Hierarchy > UI/Event System`) in order to interact with the demo UI. Also note that the UI was designed for a 16:9 game window.
 
@@ -158,7 +158,7 @@ public void RemoveItem(Item item) {
 ```
 
 ## Saving and Loading
-Saving and loading has successfully been done using a `BinaryFormatter` approach *(refer to [`BinaryIO`](/SchwerScripts/Common/NonMono/BinaryIO.cs))*. Ensure that the object you serialize uses `SerializableInventory` and not `Inventory`.
+Saving and loading has successfully been done using a `BinaryFormatter` approach *(refer to [`BinaryIO`](/schwer-scripts/Common/NonMono/BinaryIO.cs))*. Ensure that the object you serialize uses `SerializableInventory` and not `Inventory`.
 ##### Example code:
 ```csharp
 using Schwer.ItemSystem;
