+++ schwer-scripts/Common/DDOLSingleton.cs
