+++ schwer-scripts/Common/MonoBehaviourSingleton.cs
