+++ schwer-scripts/Common/NonMono/BinaryIO.cs
