+++ schwer-scripts/Editor/README.md
@@ -31,7 +31,7 @@ Editor script containing wrapper functions for working with assets.
 ## Methods
 ### `FindAllInstances<T>`
 Returns an array of all assets of a specified type in the project.
-#### Example usage (from [ItemDatabaseUtility.cs](/SchwerScripts/ItemSystem/Editor/ItemDatabaseUtility.cs)):
+#### Example usage (from [ItemDatabaseUtility.cs](/schwer-scripts/ItemSystem/Editor/ItemDatabaseUtility.cs)):
 ```csharp
 // Return a list of Item instances, omitting those with duplicate ids.
 private static List<Item> GetAllItemAssets() {
@@ -67,7 +67,7 @@ Editor script intended for working with Scriptable Object assets through code.
 ## Methods
 ### `CreateAsset<T>`
 Creates a Scriptable Object of type `T` in a process similar to `[CreateAssetMenu]`. 
-#### Example usage (from [ItemDatabaseUtility.cs](/SchwerScripts/ItemSystem/Editor/ItemDatabaseUtility.cs)):
+#### Example usage (from [ItemDatabaseUtility.cs](/schwer-scripts/ItemSystem/Editor/ItemDatabaseUtility.cs)):
 ```csharp
 private static ItemDatabase GetItemDatabase() {
     var databases = AssetsUtility.FindAllAssets<ItemDatabase>();
