+++ schwer-scripts/Editor/ScriptableObjectUtility.cs
