+++ schwer-scripts/ItemSystem/Demo/Sprites/apple.png
