+++ schwer-scripts/ItemSystem/Demo/Sprites/bow.png
