+++ schwer-scripts/ItemSystem/Demo/Sprites/shield.png
