+++ schwer-scripts/ItemSystem/Demo/Sprites/sword.png
