+++ schwer-scripts/ItemSystem/Demo/UI/InventoryManager.cs.meta
