+++ schwer-scripts/ItemSystem/Demo/UI/ItemSlot.cs.meta
