+++ schwer-scripts/ItemSystem/README.md
@@ -27,8 +27,8 @@ A skeleton implementation of items and (serializable!) inventories using Scripta
 * As `Inventory` assets are Scriptable Objects, there must be a reference to each `Inventory` asset at all times during runtime in order for their values to persist, or Unity will unload them the next time it unloads unused resources.
 
 ## Getting started
-Download this repository and add the `SchwerScripts` folder to your Unity project.
-Familiarise yourself with the system by dragging the prefab `Inventory` (`SchwerScripts/ItemSystem/Demo/UI/Prefabs/Inventory`) into a new Scene and following along with this section.
+Download this repository and add the `schwer-scripts` folder to your Unity project.
+Familiarise yourself with the system by dragging the prefab `Inventory` (`schwer-scripts/ItemSystem/Demo/UI/Prefabs/Inventory`) into a new Scene and following along with this section.
 
 You may need to create an Event System (`Right-click in the Hierarchy > UI/Event System`) in order to interact with the demo UI. Also note that the UI was designed for a 16:9 game window.
 
@@ -134,7 +134,7 @@ public void RemoveItem(Item item) {
 ```
 
 ## Saving and Loading
-Saving and loading has successfully been done using a `BinaryFormatter` approach *(refer to [`BinaryIO`](/SchwerScripts/Common/NonMono/BinaryIO.cs))*. Ensure that the object you serialize uses `SerializableInventory` and not `Inventory`.
+Saving and loading has successfully been done using a `BinaryFormatter` approach *(refer to [`BinaryIO`](/schwer-scripts/Common/NonMono/BinaryIO.cs))*. Ensure that the object you serialize uses `SerializableInventory` and not `Inventory`.
 ##### Example code:
 ```csharp
 using Schwer.ItemSystem;
